Refactor of srdf
