+++ srdf/src/literal.rs
@@ -1,4 +1,7 @@
-use std::{fmt::Display, result};
+use std::fmt::Debug;
+use std::fmt::Display;
+use std::hash::Hash;
+use std::result;
 
 use iri_s::IriS;
 use rust_decimal::{prelude::ToPrimitive, Decimal};
@@ -7,8 +10,54 @@ use serde::{Deserialize, Serialize, Serializer};
 use crate::{lang::Lang, numeric_literal::NumericLiteral};
 use prefixmap::{Deref, DerefError, IriRef, PrefixMap};
 
+pub trait Literal: Debug + Clone + Display + PartialEq + Eq + Hash {
+    fn lexical_form(&self) -> &str;
+
+    fn lang(&self) -> Option<&str>;
+
+    fn datatype(&self) -> &str;
+
+    fn as_bool(&self) -> Option<bool> {
+        match self.lexical_form() {
+            "true" => Some(true),
+            "false" => Some(false),
+            _ => None,
+        }
+    }
+
+    fn as_integer(&self) -> Option<isize> {
+        self.lexical_form().parse().ok()
+    }
+
+    fn as_double(&self) -> Option<f64> {
+        self.lexical_form().parse().ok()
+    }
+
+    fn as_decimal(&self) -> Option<Decimal> {
+        self.lexical_form().parse().ok()
+    }
+
+    fn as_literal(&self) -> SLiteral {
+        if let Some(bool) = self.as_bool() {
+            SLiteral::boolean(bool)
+        } else if let Some(int) = self.as_integer() {
+            SLiteral::integer(int)
+        } else if let Some(decimal) = self.as_double() {
+            SLiteral::double(decimal)
+        } else if let Some(decimal) = self.as_decimal() {
+            SLiteral::decimal(decimal)
+        } else if let Some(lang) = self.lang() {
+            SLiteral::lang_str(self.lexical_form(), Lang::new_unchecked(lang))
+        } else {
+            SLiteral::str(self.lexical_form())
+        }
+    }
+}
+
+/// Concrete representation of RDF literals
+/// This representation internally uses integers or doubles to represent numeric values
 #[derive(PartialEq, Eq, Hash, Debug, Serialize, Deserialize, Clone)]
-pub enum Literal {
+pub enum SLiteral {
     StringLiteral {
         lexical_form: String,
         lang: Option<Lang>,
@@ -23,39 +72,39 @@ pub enum Literal {
     BooleanLiteral(bool),
 }
 
-impl Literal {
-    pub fn integer(n: isize) -> Literal {
-        Literal::NumericLiteral(NumericLiteral::integer(n))
+impl SLiteral {
+    pub fn integer(n: isize) -> SLiteral {
+        SLiteral::NumericLiteral(NumericLiteral::integer(n))
     }
 
-    pub fn double(d: f64) -> Literal {
-        Literal::NumericLiteral(NumericLiteral::double(d))
+    pub fn double(d: f64) -> SLiteral {
+        SLiteral::NumericLiteral(NumericLiteral::double(d))
     }
 
-    pub fn decimal(d: Decimal) -> Literal {
-        Literal::NumericLiteral(NumericLiteral::decimal(d))
+    pub fn decimal(d: Decimal) -> SLiteral {
+        SLiteral::NumericLiteral(NumericLiteral::decimal(d))
     }
 
-    pub fn lit_datatype(lexical_form: &str, datatype: &IriRef) -> Literal {
-        Literal::DatatypeLiteral {
+    pub fn lit_datatype(lexical_form: &str, datatype: &IriRef) -> SLiteral {
+        SLiteral::DatatypeLiteral {
             lexical_form: lexical_form.to_owned(),
             datatype: datatype.clone(),
         }
     }
 
-    pub fn boolean(b: bool) -> Literal {
-        Literal::BooleanLiteral(b)
+    pub fn boolean(b: bool) -> SLiteral {
+        SLiteral::BooleanLiteral(b)
     }
 
-    pub fn str(lexical_form: &str) -> Literal {
-        Literal::StringLiteral {
+    pub fn str(lexical_form: &str) -> SLiteral {
+        SLiteral::StringLiteral {
             lexical_form: lexical_form.to_owned(),
             lang: None,
         }
     }
 
-    pub fn lang_str(lexical_form: &str, lang: Lang) -> Literal {
-        Literal::StringLiteral {
+    pub fn lang_str(lexical_form: &str, lang: Lang) -> SLiteral {
+        SLiteral::StringLiteral {
             lexical_form: lexical_form.to_owned(),
             lang: Some(lang),
         }
@@ -63,11 +112,11 @@ impl Literal {
 
     pub fn lexical_form(&self) -> String {
         match self {
-            Literal::StringLiteral { lexical_form, .. } => lexical_form.clone(),
-            Literal::DatatypeLiteral { lexical_form, .. } => lexical_form.clone(),
-            Literal::NumericLiteral(nl) => nl.lexical_form(),
-            Literal::BooleanLiteral(true) => "true".to_string(),
-            Literal::BooleanLiteral(false) => "false".to_string(),
+            SLiteral::StringLiteral { lexical_form, .. } => lexical_form.clone(),
+            SLiteral::DatatypeLiteral { lexical_form, .. } => lexical_form.clone(),
+            SLiteral::NumericLiteral(nl) => nl.lexical_form(),
+            SLiteral::BooleanLiteral(true) => "true".to_string(),
+            SLiteral::BooleanLiteral(false) => "false".to_string(),
         }
     }
 
@@ -77,15 +126,15 @@ impl Literal {
         prefixmap: &PrefixMap,
     ) -> std::fmt::Result {
         match self {
-            Literal::StringLiteral {
+            SLiteral::StringLiteral {
                 lexical_form,
                 lang: None,
             } => write!(f, "\"{lexical_form}\""),
-            Literal::StringLiteral {
+            SLiteral::StringLiteral {
                 lexical_form,
                 lang: Some(lang),
             } => write!(f, "\"{lexical_form}\"{lang}"),
-            Literal::DatatypeLiteral {
+            SLiteral::DatatypeLiteral {
                 lexical_form,
                 datatype,
             } => match datatype {
@@ -94,37 +143,37 @@ impl Literal {
                     write!(f, "\"{lexical_form}\"^^{}:{}", prefix, local)
                 }
             },
-            Literal::NumericLiteral(n) => write!(f, "{}", n),
-            Literal::BooleanLiteral(true) => write!(f, "true"),
-            Literal::BooleanLiteral(false) => write!(f, "false"),
+            SLiteral::NumericLiteral(n) => write!(f, "{}", n),
+            SLiteral::BooleanLiteral(true) => write!(f, "true"),
+            SLiteral::BooleanLiteral(false) => write!(f, "false"),
         }
     }
 
     pub fn datatype(&self) -> IriRef {
         match self {
-            Literal::DatatypeLiteral { datatype, .. } => datatype.clone(),
-            Literal::StringLiteral {
+            SLiteral::DatatypeLiteral { datatype, .. } => datatype.clone(),
+            SLiteral::StringLiteral {
                 lexical_form: _,
                 lang: None,
             } => IriRef::iri(IriS::new_unchecked(
                 "http://www.w3.org/2001/XMLSchema#string",
             )),
-            Literal::StringLiteral {
+            SLiteral::StringLiteral {
                 lexical_form: _,
                 lang: Some(_),
             } => IriRef::iri(IriS::new_unchecked(
                 "http://www.w3.org/1999/02/22-rdf-syntax-ns#langString",
             )),
-            Literal::NumericLiteral(NumericLiteral::Integer(_)) => IriRef::iri(
+            SLiteral::NumericLiteral(NumericLiteral::Integer(_)) => IriRef::iri(
                 IriS::new_unchecked("http://www.w3.org/2001/XMLSchema#integer"),
             ),
-            Literal::NumericLiteral(NumericLiteral::Decimal(_)) => IriRef::iri(
+            SLiteral::NumericLiteral(NumericLiteral::Decimal(_)) => IriRef::iri(
                 IriS::new_unchecked("http://www.w3.org/2001/XMLSchema#decimal"),
             ),
-            Literal::NumericLiteral(NumericLiteral::Double(_)) => IriRef::iri(IriS::new_unchecked(
-                "http://www.w3.org/2001/XMLSchema#double",
-            )),
-            Literal::BooleanLiteral(_) => IriRef::iri(IriS::new_unchecked(
+            SLiteral::NumericLiteral(NumericLiteral::Double(_)) => IriRef::iri(
+                IriS::new_unchecked("http://www.w3.org/2001/XMLSchema#double"),
+            ),
+            SLiteral::BooleanLiteral(_) => IriRef::iri(IriS::new_unchecked(
                 "http://www.w3.org/2001/XMLSchema#boolean",
             )),
         }
@@ -132,39 +181,39 @@ impl Literal {
 
     pub fn numeric_value(&self) -> Option<NumericLiteral> {
         match self {
-            Literal::NumericLiteral(nl) => Some(nl.clone()),
-            Literal::StringLiteral { .. }
-            | Literal::DatatypeLiteral { .. }
-            | Literal::BooleanLiteral(true)
-            | Literal::BooleanLiteral(false) => None,
+            SLiteral::NumericLiteral(nl) => Some(nl.clone()),
+            SLiteral::StringLiteral { .. }
+            | SLiteral::DatatypeLiteral { .. }
+            | SLiteral::BooleanLiteral(true)
+            | SLiteral::BooleanLiteral(false) => None,
         }
     }
 }
 
-impl Default for Literal {
+impl Default for SLiteral {
     fn default() -> Self {
-        Literal::StringLiteral {
+        SLiteral::StringLiteral {
             lexical_form: String::default(),
             lang: None,
         }
     }
 }
 
-impl PartialOrd for Literal {
+impl PartialOrd for SLiteral {
     fn partial_cmp(&self, other: &Self) -> Option<std::cmp::Ordering> {
         match self {
-            Literal::StringLiteral { lexical_form, .. } => match other {
-                Literal::StringLiteral {
+            SLiteral::StringLiteral { lexical_form, .. } => match other {
+                SLiteral::StringLiteral {
                     lexical_form: other_lexical_form,
                     ..
                 } => Some(lexical_form.cmp(other_lexical_form)),
                 _ => None,
             },
-            Literal::DatatypeLiteral {
+            SLiteral::DatatypeLiteral {
                 lexical_form,
                 datatype,
             } => match other {
-                Literal::DatatypeLiteral {
+                SLiteral::DatatypeLiteral {
                     lexical_form: other_lexical_form,
                     datatype: other_datatype,
                 } => {
@@ -176,19 +225,19 @@ impl PartialOrd for Literal {
                 }
                 _ => None,
             },
-            Literal::NumericLiteral(nl) => match other {
-                Literal::NumericLiteral(other_nl) => nl.partial_cmp(other_nl),
+            SLiteral::NumericLiteral(nl) => match other {
+                SLiteral::NumericLiteral(other_nl) => nl.partial_cmp(other_nl),
                 _ => None,
             },
-            Literal::BooleanLiteral(b) => match other {
-                Literal::BooleanLiteral(other_b) => Some(b.cmp(other_b)),
+            SLiteral::BooleanLiteral(b) => match other {
+                SLiteral::BooleanLiteral(other_b) => Some(b.cmp(other_b)),
                 _ => None,
             },
         }
     }
 }
 
-impl Display for Literal {
+impl Display for SLiteral {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         self.display_qualified(f, &PrefixMap::basic())
     }
@@ -204,25 +253,25 @@ where
     }
 }
 
-impl Deref for Literal {
+impl Deref for SLiteral {
     fn deref(
         &self,
         base: &Option<iri_s::IriS>,
         prefixmap: &Option<prefixmap::PrefixMap>,
     ) -> Result<Self, DerefError> {
         match self {
-            Literal::NumericLiteral(n) => Ok(Literal::NumericLiteral(n.clone())),
-            Literal::BooleanLiteral(b) => Ok(Literal::BooleanLiteral(*b)),
-            Literal::StringLiteral { lexical_form, lang } => Ok(Literal::StringLiteral {
+            SLiteral::NumericLiteral(n) => Ok(SLiteral::NumericLiteral(n.clone())),
+            SLiteral::BooleanLiteral(b) => Ok(SLiteral::BooleanLiteral(*b)),
+            SLiteral::StringLiteral { lexical_form, lang } => Ok(SLiteral::StringLiteral {
                 lexical_form: lexical_form.clone(),
                 lang: lang.clone(),
             }),
-            Literal::DatatypeLiteral {
+            SLiteral::DatatypeLiteral {
                 lexical_form,
                 datatype,
             } => {
                 let dt = datatype.deref(base, prefixmap)?;
-                Ok(Literal::DatatypeLiteral {
+                Ok(SLiteral::DatatypeLiteral {
                     lexical_form: lexical_form.clone(),
                     datatype: dt,
                 })
@@ -231,31 +280,31 @@ impl Deref for Literal {
     }
 }
 
-impl From<oxrdf::Literal> for Literal {
+impl From<oxrdf::Literal> for SLiteral {
     fn from(value: oxrdf::Literal) -> Self {
         match value.destruct() {
-            (s, None, None) => Literal::str(&s),
-            (s, None, Some(language)) => Literal::lang_str(&s, Lang::new_unchecked(&language)),
+            (s, None, None) => SLiteral::str(&s),
+            (s, None, Some(language)) => SLiteral::lang_str(&s, Lang::new_unchecked(&language)),
             (value, Some(dtype), None) => {
                 let datatype = IriRef::iri(IriS::new_unchecked(dtype.as_str()));
-                Literal::lit_datatype(&value, &datatype)
+                SLiteral::lit_datatype(&value, &datatype)
             }
             _ => todo!(),
         }
     }
 }
 
-impl From<Literal> for oxrdf::Literal {
-    fn from(value: Literal) -> Self {
+impl From<SLiteral> for oxrdf::Literal {
+    fn from(value: SLiteral) -> Self {
         match value {
-            Literal::StringLiteral { lexical_form, lang } => match lang {
+            SLiteral::StringLiteral { lexical_form, lang } => match lang {
                 Some(lang) => oxrdf::Literal::new_language_tagged_literal_unchecked(
                     lexical_form,
                     lang.to_string(),
                 ),
                 None => lexical_form.into(),
             },
-            Literal::DatatypeLiteral {
+            SLiteral::DatatypeLiteral {
                 lexical_form,
                 datatype,
             } => match datatype.get_iri() {
@@ -265,7 +314,7 @@ impl From<Literal> for oxrdf::Literal {
                 ),
                 Err(_) => lexical_form.into(),
             },
-            Literal::NumericLiteral(number) => match number {
+            SLiteral::NumericLiteral(number) => match number {
                 NumericLiteral::Integer(int) => (int as i64).into(),
                 NumericLiteral::Decimal(decimal) => match decimal.to_f64() {
                     Some(decimal) => decimal.into(),
@@ -273,7 +322,7 @@ impl From<Literal> for oxrdf::Literal {
                 },
                 NumericLiteral::Double(double) => double.into(),
             },
-            Literal::BooleanLiteral(bool) => bool.into(),
+            SLiteral::BooleanLiteral(bool) => bool.into(),
         }
     }
 }
