+++ srdf/src/srdf_builder.rs
@@ -3,10 +3,10 @@ use std::io::Write;
 use iri_s::IriS;
 use prefixmap::PrefixMap;
 
-use crate::{Query, RDFFormat};
+use crate::{NeighsRDF, RDFFormat};
 
 /// Types that implement this trait can build RDF data
-pub trait SRDFBuilder: Query {
+pub trait BuildRDF: NeighsRDF {
     /// Returns an empty RDF graph
     fn empty() -> Self;
