+++ srdf/src/srdf_parser/focus_rdf.rs
@@ -1,9 +1,9 @@
-use crate::{Query, RDFParseError};
+use crate::{NeighsRDF, RDFParseError};
 
 /// Represents RDF graphs that contain a focus node
 ///
 /// The trait contains methods to get the focus node and to set its value
-pub trait FocusRDF: Query {
+pub trait FocusRDF: NeighsRDF {
     /// Set the value of the focus node
     fn set_focus(&mut self, focus: &Self::Term);
 
@@ -24,10 +24,9 @@ pub trait FocusRDF: Query {
             None => Err(RDFParseError::NoFocusNode),
             Some(term) => {
                 let subject =
-                    Self::term_as_subject(term)
-                        .map_err(|_| RDFParseError::ExpectedSubject {
-                            node: format!("{term}"),
-                        })?;
+                    Self::term_as_subject(term).map_err(|_| RDFParseError::ExpectedSubject {
+                        node: format!("{term}"),
+                    })?;
                 Ok(subject)
             }
         }
